requirements.txt
Django
requests
Pillow
boto3
botocore
groq
psycopg2-binary
dotenv